# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Updated workspace profile
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
                  event:
                    type: object
        "403":
          description: Workspace manager permission required
    delete:
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "204":
          description: Workspace metadata deleted; queued upload-object cleanup may continue asynchronously
        "403":
          description: Workspace owner permission required
...
  /api/workspaces/{workspace_id}/github-app:
    get:
      operationId: getWorkspaceGitHubApp
      summary: Get GitHub App installations and repositories available to a workspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: GitHub App connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitHubAppStatus"
  /api/workspaces/{workspace_id}/github-app/install:
    get:
      operationId: startGitHubAppInstall
      summary: Start installing the configured GitHub App for a workspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "302":
          description: Redirect to the GitHub App installation page
        "403":
          description: Workspace manager permission required
...
  /api/hooks/github/app:
    post:
      operationId: receiveGitHubAppWebhook
      summary: Receive signed repository activity from the configured GitHub App
      responses:
        "202":
          description: Delivery accepted, ignored, or already processed
        "401":
          description: Invalid webhook signature
...
components:
  parameters:
    workspace_id:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    UpdateWorkspaceRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
...
    Workspace:
      type: object
      required: [id, route_id, name, slug, icon_url, created_at]
      properties:
        id:
          type: string
        route_id:
          type: string
          description: Immutable public route ID used in app URLs.
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
        created_at:
          type: string
          format: date-time
        role:
          type: string
          enum: [owner, moderator, member, guest, bot]
...
    GitHubAppInstallation:
      type: object
      required:
        [installation_id, workspace_id, account_login, account_type, repository_selection, installed_by, created_at, updated_at]
      properties:
        installation_id:
          type: integer
          format: int64
        workspace_id:
          type: string
        account_login:
          type: string
        account_type:
          type: string
        repository_selection:
          type: string
          enum: [all, selected]
        installed_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GitHubAppRepository:
      type: object
      required: [installation_id, id, full_name, html_url, private]
      properties:
        installation_id:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        full_name:
          type: string
        html_url:
          type: string
          format: uri
        private:
          type: boolean
    GitHubAppStatus:
      type: object
      required: [configured, installations, repositories]
      properties:
        configured:
          type: boolean
        slug:
          type: string
        installations:
          type: array
          items:
            $ref: "#/components/schemas/GitHubAppInstallation"
        repositories:
          type: array
          items:
            $ref: "#/components/schemas/GitHubAppRepository"
...